🪲 fix type conversion errors

+++ src/control/Scheduler.cpp
@@ -25,7 +25,7 @@ Scheduler::~Scheduler()
 		processors.Remove(i);
 	}
 
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		delete TRMList.peekFront();
 		TRMList.dequeue();
@@ -42,13 +42,13 @@ int Scheduler::GetCurrentTime()
 /// ////////////////////////////////// ///
 void Scheduler::CreateAllProcessors()
 {
-	for (int i = 0; i < simulationParameters.N_FCFS; i++)
+	for (unsigned int i = 0; i < simulationParameters.N_FCFS; i++)
 		CreateProcessor(FCFS);
 
-	for (int i = 0; i < simulationParameters.N_SJF; i++)
+	for (unsigned int i = 0; i < simulationParameters.N_SJF; i++)
 		CreateProcessor(SJF);
 
-	for (int i = 0; i < simulationParameters.N_RR; i++)
+	for (unsigned int i = 0; i < simulationParameters.N_RR; i++)
 		CreateProcessor(RR);
 }
 
@@ -312,7 +312,7 @@ bool Scheduler::MigrateRR(Process* process)
 bool Scheduler::MigrateFCFS(Process* process)
 {
 	TimeInfo timeInfo = process->GetTimeInfo();
-	int waitingTime = (clk->GetTime() - timeInfo.AT) - (timeInfo.CT - timeInfo.RCT);
+	unsigned int waitingTime = (clk->GetTime() - timeInfo.AT) - (timeInfo.CT - timeInfo.RCT);
 
 	if (waitingTime > simulationParameters.MAX_WAITING_TIME)
 	{
@@ -339,7 +339,7 @@ void Scheduler::ForkHandler(Process* process)
 	if (simulationParameters.N_FCFS == 0 || process == nullptr)
 		return;
 
-	int Rand = rand() % 100;
+	unsigned int Rand = rand() % 100;
 	if (Rand < simulationParameters.FORK_PROBABILITY)
 	{
 		int id = simulationParameters.N_PROCESS + 10;
@@ -687,7 +687,7 @@ std::string Scheduler::TRMListStatsToString()
 
 	unsigned int numOfTrmListProcess = TRMList.getSize();
 
-	for (int i = 0; i < numOfTrmListProcess; i++)
+	for (unsigned int i = 0; i < numOfTrmListProcess; i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -752,7 +752,7 @@ unsigned int Scheduler::CalculateAverageWaitTime()
 {
 	unsigned int totalWaitingTime = 0;
 
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -766,7 +766,7 @@ unsigned int Scheduler::CalculateAverageWaitTime()
 unsigned int Scheduler::CalculateTotalTurnaroundTime()
 {
 	unsigned int totalTurnaroundTime = 0;
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -779,7 +779,7 @@ unsigned int Scheduler::CalculateTotalTurnaroundTime()
 unsigned int Scheduler::CalculateAverageTurnaroundTime()
 {
 	unsigned int totalTurnaroundTime = 0;
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -793,7 +793,7 @@ unsigned int Scheduler::CalculateAverageResponseTime()
 {
 	unsigned int totalResponseTime = 0;
 
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -814,10 +814,10 @@ unsigned int* Scheduler::CalculateProcessorsUtilization()
 	unsigned int* processorsUtilization = new unsigned int[numOfProcessors];
 
 	// calculate utilization for each processor
-	for (int i = 0; i < numOfProcessors; i++)
+	for (unsigned int i = 0; i < numOfProcessors; i++)
 	{
 		Processor* processor = processors.GetEntry(i + 1);
-		processorsUtilization[i] = (processor->GetTotalBusyTime() / (double)totalCPUTime) * 100;
+		processorsUtilization[i] = (unsigned int)((processor->GetTotalBusyTime() / (double)totalCPUTime) * 100);
 	}
 
 	return processorsUtilization;
@@ -833,10 +833,10 @@ unsigned int* Scheduler::CalculateProcessorsLoad()
 	unsigned int* processorsLoad = new unsigned int[numOfProcessors];
 
 	// calculate load for each processor
-	for (int i = 0; i < numOfProcessors; i++)
+	for (unsigned int i = 0; i < numOfProcessors; i++)
 	{
 		Processor* processor = processors.GetEntry(i + 1);
-		processorsLoad[i] = (processor->GetTotalBusyTime() / (double)totalTurnaroundTime) * 100;
+		processorsLoad[i] = (unsigned int)((processor->GetTotalBusyTime() / (double)totalTurnaroundTime) * 100);
 	}
 
 	return processorsLoad;
@@ -850,7 +850,7 @@ unsigned int Scheduler::CalculateAverageProcessorsUtilization()
 	// calculate average utilization
 	unsigned int totalUtilization = 0;
 	unsigned int numOfProcessors = simulationParameters.N_FCFS + simulationParameters.N_SJF + simulationParameters.N_RR;
-	for (int i = 0; i < numOfProcessors; i++)
+	for (unsigned int i = 0; i < numOfProcessors; i++)
 	{
 		totalUtilization += processorsUtilization[i];
 	}
@@ -864,7 +864,7 @@ unsigned int Scheduler::CaculateWorkStealPercent()
 {
 	// count stolen processes
 	unsigned int numOfStolenProcesses = 0;
-	for (int i = 0; i < TRMList.getSize(); i++)
+	for (unsigned int i = 0; i < TRMList.getSize(); i++)
 	{
 		Process* process = TRMList.peekFront();
 		TRMList.dequeue();
@@ -875,7 +875,7 @@ unsigned int Scheduler::CaculateWorkStealPercent()
 		TRMList.enqueue(process);
 	}
 
-	return (numOfStolenProcesses / (double)TRMList.getSize()) * 100;
+	return (unsigned int)((numOfStolenProcesses / (double)TRMList.getSize()) * 100);
 }
 
 unsigned int Scheduler::GetNumberOfRTFMigrations()
@@ -890,16 +890,16 @@ unsigned int Scheduler::GetNumberOfMaxWMigrations()
 
 unsigned int Scheduler::CalculateMaxWMigrationPercent()
 {
-	return maxWMigrations * 100.f / simulationParameters.N_PROCESS;
+	return (unsigned int)(maxWMigrations * 100.f / simulationParameters.N_PROCESS);
 }
 
 unsigned int Scheduler::CalculateRTFMigrationPercent()
 {
-	return rtfMigrations * 100.f / simulationParameters.N_PROCESS;
+	return (unsigned int)(rtfMigrations * 100.f / simulationParameters.N_PROCESS);
 }
 
 unsigned int Scheduler::CalculateKillCountPercent()
 {
-	return killCount * 100.f / simulationParameters.N_PROCESS;
+	return (unsigned int)(killCount * 100.f / simulationParameters.N_PROCESS);
 }
